Make withdrawal form
This snippet is actually a shortcode [aff_withdraw]. When placed in a page, a withdrawal form showing a user available Affiliate earnings and input field will come out for Affiliate to enter how much of the earnings they will like to withdraw.  You will need to enter the minimum threshold in the snippet but I will write an update soon where you can enter it with the shortcode.